feat: new project
-Store demo project elements in backup components. -Start adding elements for the actual project including a basic first floor block on home page.
