+++ client/src/pages/Dashboard.tsx
@@ -1,6 +1,9 @@
 export default function Dashboard() {
   return (
     <div className="flex flex-col items-center justify-center h-screen bg-gray-100">
+      <nav>
+        
+      </nav>
       <h1 className="text-4xl font-bold mb-4">Dashboard</h1>
     </div>
   );
